+++ server/src-lib/Hasura/Server/API/Metadata.hs
@@ -157,7 +157,7 @@ data RQLMetadataV1
   | RMRemoveMetricsConfig
 
   -- Inherited roles
-  | RMAddInheritedRole  !AddInheritedRole
+  | RMAddInheritedRole !InheritedRole
   | RMDropInheritedRole !DropInheritedRole
 
   -- Metadata management